refactor(#255): extract registry.rs (KNOWN_SUBCOMMANDS, wrapper_targets) Move KNOWN_SUBCOMMANDS, META_SUBCOMMANDS constants, is_known_subcommand(), is_meta_subcommand(), WRAPPER_TARGETS LazyLock, and wrapper_targets() from mod.rs into a new crates/rskim/src/cmd/registry.rs module. Re-export KNOWN_SUBCOMMANDS, is_known_subcommand, is_meta_subcommand, and wrapper_targets from mod.rs. main.rs uses cmd::is_meta_subcommand() — continues to work via re-export without modification. Moves META/KNOWN sync-guard tests to registry.rs::tests. dispatch() sync-guard test remains in mod.rs (tests dispatch routing logic, not registry data). Co-Authored-By: Claude <noreply@anthropic.com>
